requirements.txt
crewai==0.28.0
crewai-tools==0.2.6
databricks-sql-connector==3.0.0
redis==5.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.4.0
sentence-transformers==2.3.1
openai==1.10.0
anthropic==0.18.0
prometheus-client==0.19.0
pyyaml==6.0.1
python-dotenv==1.0.0
